+++ hutool-core/src/test/java/cn/hutool/core/img/ImgUtilTest.java
@@ -25,7 +25,9 @@ public class ImgUtilTest {
 	@Test
 	@Ignore
 	public void scaleTest2() {
-		ImgUtil.scale(FileUtil.file("e:/pic/test.jpg"), FileUtil.file("e:/pic/test_result.jpg"), 0.8f);
+		ImgUtil.scale(
+				FileUtil.file("d:/test/2.png"),
+				FileUtil.file("d:/test/2_result.jpg"), 600, 337, null);
 	}
 
 	@Test