+++ hutool-core/src/test/java/cn/hutool/core/util/XmlUtilTest.java
@@ -9,10 +9,13 @@ import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.w3c.dom.Document;
+import org.xml.sax.Attributes;
+import org.xml.sax.helpers.DefaultHandler;
 
 import javax.xml.xpath.XPathConstants;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * {@link XmlUtil} 工具类
@@ -149,6 +152,18 @@ public class XmlUtilTest {
 	}
 
 	@Test
+	public void readBySaxTest(){
+		final Set<String> eles = CollUtil.newHashSet(
+				"returnsms", "returnstatus", "message", "remainpoint", "taskID", "successCounts");
+		XmlUtil.readBySax(ResourceUtil.getStream("test.xml"), new DefaultHandler(){
+			@Override
+			public void startElement(String uri, String localName, String qName, Attributes attributes) {
+				Assert.assertTrue(eles.contains(localName));
+			}
+		});
+	}
+
+	@Test
 	public void mapToXmlTestWithOmitXmlDeclaration() {
 
 		Map<String, Object> map = MapBuilder.create(new LinkedHashMap<String, Object>())
@@ -194,6 +209,13 @@ public class XmlUtilTest {
 		Assert.assertEquals(testBean.getOperator(), testBean2.getOperator());
 		Assert.assertEquals(testBean.getProjectCode(), testBean2.getProjectCode());
 		Assert.assertEquals(testBean.getBankCode(), testBean2.getBankCode());
+	}
+
+	@Test
+	public void cleanCommentTest() {
+		final String xmlContent = "<info><title>hutool</title><!-- 这是注释 --><lang>java</lang></info>";
+		final String ret = XmlUtil.cleanComment(xmlContent);
+		Assert.assertEquals("<info><title>hutool</title><lang>java</lang></info>", ret);
 	}
 
 	@Data